==> analysis/2_analysis.qmd <==
---
title: "FaceFake2 - Data Analysis"
editor: source
editor_options: 
  chunk_output_type: console
format:
  html:
    code-fold: true
    self-contained: true
    toc: true
---
## Summary

All models included random slopes for Condition by participant, as well as random intercepts for both participants and images. 

Notably, only the attractiveness model successfully converged when random slopes for Relevance were added at the image level. While it is generally recommended to specify the maximal random-effects structure justified by the design to improve generalizability [Bar et al., 2023]([https://doi.org/10.3389/fpsyg.2013.00328) , maintaining consistency across models is also important to enable meaningful comparisons, hence the latter was prioritised. 


The **effect of image condition** (AI-generated vs. photo) differed significantly between females and males.

For females, images labeled as AI-generated were rated lower in arousal, beauty, and trustworthiness when the images were irrelevant. When the images were relevant, AI-generated labels led to lower ratings in beauty and trustworthiness.

For males, the only significant effect of condition was that irrelevant images labeled as AI-generated were rated lower in beauty.

**Moderators** for subjective ratings accounting for the effect of **Condition** was only **significant for females**. Specifically: 

1. Females' attractiveness ratings of *irrelevant* images labelled *AI-generated* was moderated by *Honesty-Humility* ratings (see figures below)
2. Females' beauty ratings of *relevant* images labelled *AI-generated* was moderated by *Conscientiousness* ratings (see figures bellow)
3. Female's trustworthiness ratings of *irrelevant* images were moderated by *Openness to Experience*, *Honesty-Humility* and *Conscientiousness*. 
4. While, female's trsutworthiness ratings of *relevant* images were moderated by *Emotionality* and *Conscientiousness*. 

## Data Preparation

```{r}
#| message: false
#| warning: false

library(tidyverse)
library(easystats)
library(patchwork)
library(ggside)

results_table <- function(model, effects="fixed", filter=NULL) {
  if("marginaleffects" %in% class(model)) {
    model |> 
      parameters::parameters() |> 
      as.data.frame() |> 
      select(-Parameter, -SE, -S, z=Statistic) |> 
      insight::format_table() |> 
      parameters::display()
  } else {
    display(parameters::parameters(model, effects=effects, keep=filter))
  }
}
```


```{r}
#| code-fold: true

df <- read.csv("../data/data_participants.csv") 
dftask <- read.csv("../data/data_task.csv") |> 
  mutate(Condition = fct_relevel(Condition, "Photograph", "AI-Generated"),
         Relevance = fct_relevel(Relevance, "Relevant", "Irrelevant"),
         Attractiveness = Attractiveness / 6,
         Beauty = Beauty / 6,
         Trustworthiness = Trustworthiness / 6,
         Realness = Realness / 6 + 0.5,
         RealnessBelief = ifelse(Realness > 0.5, 1, 0),
         Gender = fct_relevel(Gender, "Male", "Female"))

dftask <- full_join(dftask, 
                    select(df, Participant, starts_with("HEXACO"), -ends_with("_NR"), -ends_with("_R")),
                    by="Participant")
```

## Visualisation of Variables

::: panel-tabset

### Attractiveness 

```{r}
dftask |> 
  ggplot(aes(x=Attractiveness, fill=Condition)) +
  geom_bar(aes(y = after_stat(prop)), position="dodge") +
  facet_grid(Relevance~Gender) +
  theme_bw()
  
```

This model looks at the effect of Gender and Relevance on attractiveness scores, accounting for random variabilty due to participants and items (i.e., random effects). 

Females did not rate faces significantly higher than males in attractiveness. However, both genders rated irrelevant images lower in attractiveness than relevant ones.

```{r}
m_a<- glmmTMB::glmmTMB(Attractiveness ~ Gender / Relevance + 
                                       (Relevance | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m_a)

```

```{r}
estimate_relation(m_a) |> 
  ggplot(aes(x=Relevance, y=Predicted)) +
  geom_pointrange(aes(ymin=CI_low, ymax=CI_high, color=Relevance), position=position_dodge(width=0.5)) +
  scale_color_manual(values=c("Relevant"="#03A9F4", "Irrelevant"="#FF9800"), guide = "none") +
  labs(y="Attractiveness") +
  facet_wrap(~Gender) +
  theme_bw()
```

### Beauty

```{r}
dftask |> 
  ggplot(aes(x=Beauty, fill=Condition)) +
  geom_bar(aes(y = after_stat(prop)), position="dodge") +
  facet_grid(Relevance~Gender) +
  theme_bw()
```

This model looks at the effect of Gender and Relevance on beauty scores, accounting for random variabilty due to participants and items (i.e., random effects). 

Females rated faces significantly higher than males in beauty. Males rated irrelevant images as more beautiful than relevant ones, whilst females rated irrelevant images as less beautiful than relevent ones, both effects were significant. 

```{r}
m_b<- glmmTMB::glmmTMB(Beauty ~ Gender / Relevance + 
                                       (Relevance | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m_b)

```

```{r}
estimate_relation(m_b) |> 
  ggplot(aes(x=Relevance, y=Predicted)) +
  geom_pointrange(aes(ymin=CI_low, ymax=CI_high, color=Relevance), position=position_dodge(width=0.5)) +
  scale_color_manual(values=c("Relevant"="#03A9F4", "Irrelevant"="#FF9800"), guide = "none") +
  labs(y="Beauty") +
  facet_wrap(~Gender) +
  theme_bw()
```

### Trustworthiness

```{r}
dftask |> 
  ggplot(aes(x=Trustworthiness, fill=Condition)) +
  geom_bar(aes(y = after_stat(prop)), position="dodge") +
  facet_grid(Relevance~Gender) +
  theme_bw()
```

This model looks at the effect of Gender and Relevance on trustworthiness scores, accounting for random variabilty due to participants and items (i.e., random effects). 

Females rated faces higher in trustworthiness then males, however this effect was not significant. Males rated irrelevant images as significantly more trustworthy than relevant ones, whereas the same effect was not significant for women. 

```{r}
m_t<- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance + 
                                       (Relevance | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m_t)

```

```{r}
estimate_relation(m_b) |> 
  ggplot(aes(x=Relevance, y=Predicted)) +
  geom_pointrange(aes(ymin=CI_low, ymax=CI_high, color=Relevance), position=position_dodge(width=0.5)) +
  scale_color_manual(values=c("Relevant"="#03A9F4", "Irrelevant"="#FF9800"), guide = "none") +
  labs(y="Beauty") +
  facet_wrap(~Gender) +
  theme_bw()

```

### Reality 

```{r}
dftask |> 
  ggplot(aes(x=Realness, fill=Condition)) +
  geom_bar(aes(y = after_stat(prop)), position="dodge") +
  facet_grid(Relevance~Gender) +
  theme_bw()
```

The model evaluating the effect of Gender and Relevance on Realness ratings shows no significant effects. 

```{r}
m_r<- glmmTMB::glmmTMB(Realness ~ Gender / Relevance + 
                                       (Relevance | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m_r)

```

```{r}
estimate_relation(m_r) |> 
  ggplot(aes(x=Relevance, y=Predicted)) +
  geom_pointrange(aes(ymin=CI_low, ymax=CI_high, color=Relevance), position=position_dodge(width=0.5)) +
  scale_color_manual(values=c("Relevant"="#03A9F4", "Irrelevant"="#FF9800"), guide = "none") +
  labs(y="Realness") +
  facet_wrap(~Gender) +
  theme_bw()
```

:::

## Attractiveness 

This model examines the effects of Gender, Relevance, and Condition on Attractiveness, while accounting for random variability by including random intercepts for Participants and Stimuli, as well as random slopes for Condition (by Participant) and Relevance (by Stimulus).

The model demonstrated strong conditional fit (R²_conditional = 0.812), indicating that most of the explained variance was due to random effects, particularly between participants (ICC = 0.810).

Notably, only female participants rated irrelevant images as significantly less attractive than relevant ones in the AI-generated condition. No significant effects were observed for male participants or other interaction terms. 

The variance decomposition (D_vour) further confirms that most of the explainable variance lies in participant-level intercepts (0.98) and stimulus-level intercepts (0.97), with smaller contributions from the random slopes.


```{r}
m_attractiveness <- glmmTMB::glmmTMB(Attractiveness ~ Gender / Relevance / Condition + 
                                       (Condition | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

performance::performance(m_attractiveness)
results_table(m_attractiveness, filter="Condition")
performance::performance_dvour(m_attractiveness)

```

## Beauty 

This model examines the effects of Gender, Relevance, and Condition on Attractiveness, while accounting for random variability by including random intercepts for Participants and Stimuli, as well as random slopes for Condition (by Participant).

The model demonstrated strong conditional fit (R²_conditional = 0.858), indicating that most of the explained variance was due to random effects, particularly between participants (ICC = 0.849).

Females rated relevant images significantly lower in beauty than males as well as rating irrelevant images signifcantly lower in beauty than relevant ones. Lastly, males also rated irrelevant images lower in beauty than relevant ones. 

The variance decomposition (D_vour) further declares the smaller contributions from the random slopes. 

<!-- *Note*. models with random random slopea for stimulus were singular.  -->

```{r}
m_beauty <- glmmTMB::glmmTMB(Beauty ~ Gender / Relevance / Condition + (Condition | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m_beauty, filter="Condition")
performance::performance(m_beauty)
performance::performance_dvour(m_beauty)
```

```{r}
# modelbased::estimate_grouplevel(m_beauty)
```


## Trustworthiness 

This model examines the effects of Gender, Relevance, and Condition on Attractiveness, while accounting for random variability by including random intercepts for Participants and Stimuli, as well as random slopes for Condition (by Participant).

The model demonstrated strong conditional fit (R²_conditional = 0.845), indicating that most of the explained variance was due to random effects, particularly between participants (ICC = 0.844).

Females rated relevant images significantly lower in trustworthiness than males as well as rating irrelevant images signifcantly lower in trustworthiness than relevant ones. Males did not rate irrelevant images significantly lower in trustworthiness than relevant ones. 

The variance decomposition (D_vour) further confirms that most of the explainable variance lies in participant-level intercepts (0.98) and stimulus-level intercepts (0.96).

<!-- *Note* Models with random slopes for stimuli were singular -->
```{r}
m_trustworthiness <- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance / Condition + (Condition|Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m_trustworthiness, filter="Condition")
performance::performance(m_trustworthiness)
performance::performance_dvour(m_trustworthiness)


```

## Coupling - Does fiction create a decoupling?

This model evaluates how beauty and image condition (e.g., AI-generated or not) jointly influence attractiveness ratings, and how this relationship is further moderated by the relevance of the image and the gender of the participant.

Beauty ratings are a strong predictor of attractiveness, but the strenght of this relationship varies by gender and relevance. Condition does not significantly moderate the beauty-attractiveness relationship, indicating that fiction does not create a decoupling between beauty and attractiveness.

```{r}
m_coupling1 <- glmmTMB::glmmTMB(Attractiveness ~ Gender / Relevance / (Beauty * Condition) + 
                               (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m_coupling1, filter="Beauty")
performance::performance(m_coupling1)


```

This model evaluates how beauty and image condition (e.g., AI-generated or not) jointly influence trustworthiness ratings, and how this relationship is further moderated by the relevance of the image and the gender of the participant.

Beauty ratings are a strong predictor of trustowrthiness but condition does not significantly moderate the beauty-trustworthiness relationship. These findings indicate that fiction does not create a decoupling between beauty and trustworthiness. 

```{r}
m_coupling2 <- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance / (Beauty * Condition) + 
                               (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      # family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m_coupling2, filter="Beauty")
performance::performance(m_coupling2)

```

This model evaluates how attractiveness and image condition (e.g., AI-generated or not) jointly influence trustworthiness ratings, and how this relationship is further moderated by the relevance of the image and the gender of the participant.

Atractiveness ratings are a strong predictor of trustowrthiness, but the strenght of this relationship varies by gender and relevance. 

Importantly, Condition (AI-generated vs. real) significantly moderates the attractiveness–trustworthiness relationship only for females rating irrelevant images. In this group, real photos are rated as less trustworthy compared to AI-generated images at comparable levels of attractiveness, indicating a weakening or decoupling of the usual link between attractiveness and trustworthiness when the image is irrelevant and real.

These results suggest that fiction (AI-generated images) affects the attractiveness–trustworthiness association for females evaluating irrelevant images, creating a decoupling between perceived beauty and trustworthiness in this context.
```{r}
m_coupling3 <- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance / (Attractiveness * Condition) + 
                               (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      # family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m_coupling3, filter="Attractiveness")
performance::performance(m_coupling3)

```


## Moderator

### Attractiveness 

::: panel-tabset

#### Honest-Humility

```{r}
m1 <- glmmTMB::glmmTMB(Attractiveness ~ Gender / Relevance / (Condition * HEXACO18_HonestyHumility) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m1, filter="HEXACO18_HonestyHumility")
```

#### Openness

Honesty-Humility significantly moderated females’ ratings of irrelevant images labeled as AI-generated vs photos, such that higher Honesty-Humility was associated with lower ratings attractiveness. 

```{r}
m2 <- glmmTMB::glmmTMB(Attractiveness ~ Gender / Relevance / (Condition * HEXACO18_Openness) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m2, filter="HEXACO18_Openness")
```

```{r}

predao <- estimate_relation(m2, lenght = 50)

p1 <- predao |>
  filter(Gender== "Female") |>
   ggplot(aes(x=HEXACO18_Openness , y=Predicted)) +
  geom_ribbon(aes(ymin=CI_low, ymax=CI_high, fill=Condition), alpha=0.3) +
  geom_line(aes(color=Condition), linewidth=1, key_glyph = draw_key_rect) +
  # geom_text(data=stars1, aes(x=x, y=y, label=label, color=Condition), hjust=0.5, size=3) +
  geom_text(data=data.frame(Relevance=as.factor("Irrelevant"), x=4, y=0.3156, label="**"), aes(x=x, y=y, label=label), hjust=0.5, color="#424242", size=6) +
  facet_grid(~Relevance) +
  scale_y_continuous(labels=scales::percent) +
  scale_color_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336")) +
  scale_fill_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336"), guide="none") +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 8),
        strip.placement = "outside",
        strip.background.x = element_rect(fill=c("lightgrey"), color=NA),
        strip.text.x = element_text(size = 10),
        strip.text.y = element_text(size = 10),
        axis.text.x = element_text(size=9, color="black"),
        legend.text = element_text(size = 10)) +
  labs(y="Attractiveness\n", fill="Images presented as:", color="Images presented as:",
       x = "\nOppeness to Experiences",
       caption ="Note. Females Only")
```

#### Emotionality

```{r}
m3 <- glmmTMB::glmmTMB(Attractiveness ~ Gender / Relevance / (Condition * HEXACO18_Emotionality) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m3, filter="HEXACO18_Emotionality")
```

#### Extraversion

```{r}
m4 <- glmmTMB::glmmTMB(Attractiveness ~ Gender / Relevance / (Condition * HEXACO18_Extraversion) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m4, filter="HEXACO18_Extraversion")
```

#### Agreeableness

```{r}
m5 <- glmmTMB::glmmTMB(Attractiveness ~ Gender / Relevance / (Condition * HEXACO18_Agreeableness) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m5, filter="HEXACO18_Agreeableness")
```

#### Conscientiousness

```{r}
m6 <- glmmTMB::glmmTMB(Attractiveness ~ Gender / Relevance / (Condition * HEXACO18_Conscientiousness) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m6, filter="HEXACO18_Conscientiousness")
```

:::

### Beauty 

::: panel-tabset

#### Honest-Humility

```{r}
m1 <- glmmTMB::glmmTMB(Beauty ~ Gender / Relevance / (Condition * HEXACO18_HonestyHumility) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m1, filter="HEXACO18_HonestyHumility")
```

#### Openness

```{r}
m2 <- glmmTMB::glmmTMB(Beauty ~ Gender / Relevance / (Condition * HEXACO18_Openness) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m2, filter="HEXACO18_Openness")
```

#### Emotionality

```{r}
m3 <- glmmTMB::glmmTMB(Beauty ~ Gender / Relevance / (Condition * HEXACO18_Emotionality) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m3, filter="HEXACO18_Emotionality")
```

#### Extraversion

```{r}
m4 <- glmmTMB::glmmTMB(Beauty ~ Gender / Relevance / (Condition * HEXACO18_Extraversion) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m4, filter="HEXACO18_Extraversion")
```

#### Agreeableness

```{r}
m5 <- glmmTMB::glmmTMB(Beauty ~ Gender / Relevance / (Condition * HEXACO18_Agreeableness) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m5, filter="HEXACO18_Agreeableness")
```

#### Conscientiousness

Conscientiousness significantly moderated females’ ratings of relevant images labeled as AI-generated vs photos, such that higher conscientiousness was associated with more positive ratings of beauty.

Additionally, conscientiousness significantly moderated males’ ratings of relevant images that are presented as photos. 

```{r}
m6 <- glmmTMB::glmmTMB(Beauty ~ Gender / Relevance / (Condition * HEXACO18_Conscientiousness) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m6, filter="HEXACO18_Conscientiousness")
```

```{r}

predbc <- estimate_relation(m6, lenght = 50)

p2 <- predbc |>
  filter(Gender== "Female") |>
   ggplot(aes(x=HEXACO18_Conscientiousness , y=Predicted)) +
  geom_ribbon(aes(ymin=CI_low, ymax=CI_high, fill=Condition), alpha=0.3) +
  geom_line(aes(color=Condition), linewidth=1, key_glyph = draw_key_rect) +
  # geom_text(data=stars1, aes(x=x, y=y, label=label, color=Condition), hjust=0.5, size=3) +
  geom_text(data=data.frame(Relevance=as.factor("Relevant"), x=3, y=0.493, label="*"), aes(x=x, y=y, label=label), hjust=0.5, color="#424242", size=6) +
  facet_grid(~Relevance) +
  scale_y_continuous(labels=scales::percent) +
  scale_color_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336")) +
  scale_fill_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336"), guide="none") +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 8),
        strip.placement = "outside",
        strip.background.x = element_rect(fill=c("lightgrey"), color=NA),
        strip.text.x = element_text(size = 10),
        strip.text.y = element_text(size = 10),
        axis.text.x = element_text(size=9, color="black"),
        legend.text = element_text(size = 10)) +
  labs(y="Beauty", fill="Images presented as:", color="Images presented as:",
       x = "\nConscientiousness")

```

:::

### Trustworthiness

::: panel-tabset

#### Honest-Humility

Honesty-Humility significantly moderated trustworthiness ratings for both relevant and irrelevant images in both males and females. Higher Honesty-Humility was associated with higher trustworthiness ratings across all conditions, except when images were labelled as AI-generated — where no effects were found except for females rating irrelevant AI-generated images, where higher Honesty-Humility was associated with lower trustworthiness. 


```{r}
m1 <- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance / (Condition * HEXACO18_HonestyHumility) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m1, filter="HEXACO18_HonestyHumility")
```

```{r}
predthh <- estimate_relation(m1, lenght = 50)

p3 <- predthh |>
  filter(Gender== "Female") |>
   ggplot(aes(x=HEXACO18_HonestyHumility , y=Predicted)) +
  geom_ribbon(aes(ymin=CI_low, ymax=CI_high, fill=Condition), alpha=0.3) +
  geom_line(aes(color=Condition), linewidth=1, key_glyph = draw_key_rect) +
  # geom_text(data=stars1, aes(x=x, y=y, label=label, color=Condition), hjust=0.5, size=3) +
  geom_text(data=data.frame(Relevance=as.factor("Irrelevant"), x=4, y=0.485, label="*"), aes(x=x, y=y, label=label), hjust=0.5, color="#424242", size=6) +
  facet_grid(~Relevance) +
  scale_y_continuous(labels=scales::percent) +
  scale_color_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336")) +
  scale_fill_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336"), guide="none") +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 8),
        strip.placement = "outside",
        strip.background.x = element_rect(fill=c("lightgrey"), color=NA),
        strip.text.x = element_text(size = 10),
        strip.text.y = element_text(size = 10),
        axis.text.x = element_text(size=9, color="black"),
        legend.text = element_text(size = 10)) +
  labs(y="Trustworthiness\n", fill="Images presented as:", color="Images presented as:",
       x = "\nHonesty-Humility",
       caption ="Note. Females Only")

```

#### Openness

Openness significantly moderated females’ ratings of irrelevant AI-generated images, such that higher Openness was associated with lower trustworthiness ratings. No other effects were significant.

```{r}
m2 <- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance / (Condition * HEXACO18_Openness) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m2, filter="HEXACO18_Openness")
```

```{r}

predto <- estimate_relation(m2, lenght = 50)

p4 <- predto |>
  filter(Gender== "Female") |>
   ggplot(aes(x=HEXACO18_Openness , y=Predicted)) +
  geom_ribbon(aes(ymin=CI_low, ymax=CI_high, fill=Condition), alpha=0.3) +
  geom_line(aes(color=Condition), linewidth=1, key_glyph = draw_key_rect) +
  # geom_text(data=stars1, aes(x=x, y=y, label=label, color=Condition), hjust=0.5, size=3) +
  geom_text(data=data.frame(Relevance=as.factor("Irrelevant"), x=4, y=0.485, label="*"), aes(x=x, y=y, label=label), hjust=0.5, color="#424242", size=6) +
  facet_grid(~Relevance) +
  scale_y_continuous(labels=scales::percent) +
  scale_color_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336")) +
  scale_fill_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336"), guide="none") +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 8),
        strip.placement = "outside",
        strip.background.x = element_rect(fill=c("lightgrey"), color=NA),
        strip.text.x = element_text(size = 10),
        strip.text.y = element_text(size = 10),
        axis.text.x = element_text(size=9, color="black"),
        legend.text = element_text(size = 10)) +
  labs(y="Trustworthiness\n", fill="Images presented as:", color="Images presented as:",
       x = "\nOpenness to Experience",
       caption ="Note. Females Only")
```

#### Emotionality

Emotionality significantly moderated females’ trustworthiness ratings of relevant AI-generated images, where higher Emotionality was associated with lower trustworthiness. No other significant effects were found.

```{r}
m3 <- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance / (Condition * HEXACO18_Emotionality) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m3, filter="HEXACO18_Emotionality")
```

```{r}

predte <- estimate_relation(m3, lenght = 50)

p5 <- predte |>
  filter(Gender== "Female") |>
   ggplot(aes(x=HEXACO18_Emotionality , y=Predicted)) +
  geom_ribbon(aes(ymin=CI_low, ymax=CI_high, fill=Condition), alpha=0.3) +
  geom_line(aes(color=Condition), linewidth=1, key_glyph = draw_key_rect) +
  # geom_text(data=stars1, aes(x=x, y=y, label=label, color=Condition), hjust=0.5, size=3) +
  geom_text(data=data.frame(Relevance=as.factor("Relevant"), x=4.25, y=0.485, label="*"), aes(x=x, y=y, label=label), hjust=0.5, color="#424242", size=6) +
  facet_grid(~Relevance) +
  scale_y_continuous(labels=scales::percent) +
  scale_color_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336")) +
  scale_fill_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336"), guide="none") +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 8),
        strip.placement = "outside",
        strip.background.x = element_rect(fill=c("lightgrey"), color=NA),
        strip.text.x = element_text(size = 10),
        strip.text.y = element_text(size = 10),
        axis.text.x = element_text(size=9, color="black"),
        legend.text = element_text(size = 10)) +
  labs(y="Trustworthiness\n", fill="Images presented as:", color="Images presented as:",
       x = "\nEmotionality",
       caption ="Note. Females Only")
```

#### Extraversion

Extraversion significantly moderated males’ ratings of irrelevant images, such that higher Extraversion was associated with higher trustworthiness ratings. No other effects were significant. 

```{r}
m4 <- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance / (Condition * HEXACO18_Extraversion) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m4, filter="HEXACO18_Extraversion")
```

#### Agreeableness

Agreeableness significantly moderated females’ ratings of relevant images, with higher Agreeableness being associated with higher trustworthiness ratings. Other interactions were non-significant.

```{r}
m5 <- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance / (Condition * HEXACO18_Agreeableness) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m5, filter="HEXACO18_Agreeableness")
```

#### Conscientiousness

Conscientiousness significantly moderated females’ trustworthiness ratings of AI-generated images. Specifically, higher Conscientiousness was associated with higher trustworthiness ratings for relevant AI-generated images and lower trustworthiness ratings for irrelevant AI-generated images.

```{r}
m6 <- glmmTMB::glmmTMB(Trustworthiness ~ Gender / Relevance / (Condition * HEXACO18_Conscientiousness) + 
                                       (1 | Participant) + (1|Stimulus),
                      data=dftask, 
                      family=glmmTMB::ordbeta(),
                      control = glmmTMB::glmmTMBControl(parallel = 8))

results_table(m6, filter="HEXACO18_Conscientiousness")
```

```{r}

predtc <- estimate_relation(m6, lenght = 50)

p6 <- predtc |>
  filter(Gender== "Female") |>
   ggplot(aes(x=HEXACO18_Conscientiousness , y=Predicted)) +
  geom_ribbon(aes(ymin=CI_low, ymax=CI_high, fill=Condition), alpha=0.3) +
  geom_line(aes(color=Condition), linewidth=1, key_glyph = draw_key_rect) +
  # geom_text(data=stars1, aes(x=x, y=y, label=label, color=Condition), hjust=0.5, size=3) +
  geom_text(data=data.frame(Relevance=as.factor("Relevant"), x=3, y=0.485, label="*"), aes(x=x, y=y, label=label), hjust=0.5, color="#424242", size=6) +
    geom_text(data=data.frame(Relevance=as.factor("Irrelevant"), x=3.2, y=0.485, label="*"), aes(x=x, y=y, label=label), hjust=0.5, color="#424242", size=6) +
  facet_grid(~Relevance) +
  scale_y_continuous(labels=scales::percent) +
  scale_color_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336")) +
  scale_fill_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336"), guide="none") +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 8),
        strip.placement = "outside",
        strip.background.x = element_rect(fill=c("lightgrey"), color=NA),
        strip.text.x = element_text(size = 10),
        strip.text.y = element_text(size = 10),
        axis.text.x = element_text(size=9, color="black"),
        legend.text = element_text(size = 10)) +
  labs(y="Trustworthiness\n", fill="Images presented as:", color="Images presented as:",
       x = "\nConscientiousness",
       caption ="Note. Females Only")
```

:::


## Reality 

Please refer to this [analysis on the reality determinants for this study](https://realitybending.github.io/FictionChoco/analysis/1_study2.html). Instead of using a traditional ZOIB model to examine whether realness scores were predicted by Gender and Condition, this analysis employed a novel modelling approach designed to better capture subjective scale responses. Specifically, the new model accounts for both the binary aspect of the decision (e.g., True/False, Agree/Disagree) and the intensity of that choice, quantified by how far the cursor was moved from the scale’s midpoint. This allows for a more nuanced interpretation of responses that may reflect an underlying discrete choice process.

## Figures

### Main Effects

```{r}
#| fig-height: 7
#| fig-width: 7


pred_attr <- modelbased::estimate_means(m_attractiveness, by=c("Gender", "Relevance", "Condition")) |>
  as.data.frame() |>
  mutate(Outcome="Attractiveness")

pred_beauty <- modelbased::estimate_means(m_beauty, by=c("Gender", "Relevance", "Condition")) |>
  as.data.frame() |>
  mutate(Outcome="Beauty")

pred_trust <- modelbased::estimate_means(m_trustworthiness, by=c("Gender", "Relevance", "Condition")) |>
  as.data.frame() |>
  mutate(Outcome="Trustworthiness")

pred4 <- rbind(pred_attr, pred_beauty, pred_trust)

stars1 <- data.frame(
  label = c("**", "**", "*", "***", "***", "***"),
  Gender = factor(c("Female", "Female", "Male", "Female", "Female", "Female"), levels = c("Male", "Female")),
  Relevance = c("Irrelevant", "Relevant", "Irrelevant", "Relevant", "Relevant", "Irrelevant"),
  Outcome = c("Attractiveness", "Beauty", "Beauty", "Beauty", "Trustworthiness", "Trustworthiness"),
  y = c(0.39, 0.52, 0.42, 0.55, 0.54, 0.54),
  x = c(2.0, 2, 1.0, 1, 1, 2))


p7 <- pred4 |> 
  ggplot(aes(x=Relevance, y=Proportion)) +
  # geom_hline(yintercept=0, color="darkgrey") +
  geom_line(aes(group=Condition, color=Condition), position = position_dodge(width=0.2), alpha=0.8, linewidth=0.2) +
  geom_pointrange(aes(ymin =CI_low, ymax=CI_high, color=Condition), position = position_dodge(width=0.2), size=0.2) +
  facet_grid(Outcome~Gender, switch="y", scales="free") +
  scale_y_continuous(labels=scales::percent) +
  scale_color_manual(values=c("AI-Generated"="#2196F3", "Photograph"="#F44336")) +
  geom_text(data = stars1,aes(x = x, y = y, label = label),inherit.aes = FALSE,size = 4
  ) +
  geom_rect(data=data.frame(Gender=as.factor("Male"), Proportion=0.1,  Outcome="Attractiveness", Relevance="Relevant"),
            aes(xmin=0.8, xmax=1.2, ymin=0.28, ymax=0.42), color="#8BC34A", alpha=0) +
  geom_rect(data=data.frame(Gender=as.factor("Female"), Proportion=0.1,  Outcome="Attractiveness", Relevance="Relevant"),
            aes(xmin=0.8, xmax=1.2, ymin=0.33, ymax=0.44), color="#8BC34A", alpha=0) +
  # coord_cartesian(ylim=c(0.15, 0.6)) +
  theme_bw() +
  theme(axis.title.y = element_blank(),
        axis.title.x = element_blank(),
        axis.text.y = element_text(size = 8),
        strip.placement = "outside",
        strip.background.x = element_rect(fill=c("lightgrey", "blue"), color=NA),
        strip.text.x = element_text(size = 10),
        strip.text.y = element_text(size = 10),
        axis.text.x = element_text(size=9, color="black"),
        legend.text = element_text(size = 10),
        legend.position = ) +
  labs(color="Images presented as:")

p7
ggsave("figures/figure1.png", p7, width=8, height=8, dpi=300)


```


```{r}
#| fig-height: 7
#| fig-width: 12
p8 <- patchwork::wrap_plots(p1|p2, guides="collect")
p8

```

```{r}
#| fig-height: 7
#| fig-width: 12


p9 <- patchwork::wrap_plots((p3|p4)/(p5|p6), guides="collect")
ggsave("figures/figure3.png", p9, width=10, height=10, dpi=300)
p9

```

## Notes

- The attractiveness, beauty and trustworthy model with relevance and condition as random slopes for stimuli were singular.
- The beauty and trustworthiness models with either relevance or condition as random slopes for stimuli were singular. 
- The D-vour score was higher for the models with Relevance as a random slope on Stimulus than for the model with Condition, suggesting that Relevance captures more reliable variability across stimuli. Therefore, Relevance was retained as the random slope for Stimulus (see [explanation of D-vour](https://easystats.github.io/performance/reference/performance_reliability.html)).
